feat(compaction): O_DIRECT reads and XL blob support
Compaction now uses O_DIRECT for both reads and writes to avoid page
cache pollution during segment merges. Key changes:

- Add sys.AlignRange() for computing O_DIRECT-compliant read parameters
- Add sys.OpenFileForRead() and sys.PreadAligned() for aligned reads
- Compactor caches O_DIRECT file handles for source segments
- Simplify interface: replace SegmentReader with SegmentCacheReleaseFn
- Use sys.AllocAligned directly instead of MmapBuffer (no ref counting needed)
- Handle XL blobs (larger than buffer): flush pending, alloc temp buffer,
  read/write directly, free temp buffer
- Track diskWritten separately from bufPos for correct offset calculation
- Remove ReadBlobRaw from Archivist (no longer needed)

+++ memtable.go
@@ -39,7 +39,7 @@ type MemTable struct {
 	segIDs    SegmentIDProvider
 	slabPool  *MmapPool
 	publisher Publisher
-	wal        *wal.WAL // nil if WAL disabled
+	wal       *wal.WAL // nil if WAL disabled
 
 	mu struct {
 		sync.Mutex
@@ -239,7 +239,7 @@ func (mt *MemTable) writeToSlab(seqID uint64, hash Key, rec record.Record) error
 			// Reserve position and increment xlSize under lock.
 			// Actual buffer allocation happens after unlock to avoid mmap syscall under lock.
 			wPos = active.AlignPosToPageBoundary()
-			active.xlSize += int64(sys.PageAlign(record.FileHeaderSize + int64(writeSize)))
+			active.xlSize += sys.PageAlign(record.FileHeaderSize + int64(writeSize))
 		} else {
 			needRotation = true
 		}
